Тинькофф/Проекты/Распознавание лиц/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deepface==0.0.79
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gdown==4.7.1
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.9.1.0
mtcnn==0.1.1
numpy==1.25.0
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==3.15.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.65.0
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
